index.html
The Tower of Hanoi is a classic mathematical puzzle that challenges your strategic thinking! The objective is to move an entire stack of disks from one peg to another, following three simple rules:

1.  Only one disk can be moved at a time.
2.  Each move consists of taking the upper disk from one of the stacks and placing it on top of another stack or on an empty peg.
3.  No disk may be placed on top of a smaller disk.

Our interactive version allows you to play with a customizable number of disks (from 3 to 7), track your moves, and even see an auto-solve demonstration. You can move disks by clicking to select and then clicking a target peg, or by using the intuitive drag-and-drop functionality.